Split fakes and fixtures into separate files
