+++ readium/lcp/src/test/java/org/readium/r2/lcp/LcpContentProtectionServiceTest.kt
@@ -6,67 +6,24 @@
 
 package org.readium.r2.lcp
 
-import kotlin.time.Instant
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import org.json.JSONObject
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.readium.r2.lcp.license.model.LicenseDocument
-import org.readium.r2.lcp.license.model.StatusDocument
+import org.readium.r2.lcp.fakes.FakeLcpLicense
 import org.readium.r2.shared.publication.LocalizedString
 import org.readium.r2.shared.publication.Manifest
 import org.readium.r2.shared.publication.Metadata
 import org.readium.r2.shared.publication.Publication
 import org.readium.r2.shared.publication.protection.ContentProtection
 import org.readium.r2.shared.publication.services.ContentProtectionService
-import org.readium.r2.shared.util.Try
 import org.readium.r2.shared.util.data.EmptyContainer
 import org.robolectric.RobolectricTestRunner
 
 @RunWith(RobolectricTestRunner::class)
 class LcpContentProtectionServiceTest {
-
-    private class FakeLcpLicense(
-        override val canCopy: Boolean = true,
-        override val canPrint: Boolean = true,
-    ) : LcpLicense {
-        var closed = false
-
-        override val license: LicenseDocument
-            get() = LicenseDocument(json = JSONObject())
-        override val status: StatusDocument? = null
-        override val charactersToCopyLeft: StateFlow<Int?> = MutableStateFlow(null)
-        override val pagesToPrintLeft: StateFlow<Int?> = MutableStateFlow(null)
-        override val canRenewLoan: Boolean = false
-        override val maxRenewDate: Instant? = null
-        override val canReturnPublication: Boolean = false
-
-        override suspend fun renewLoan(
-            listener: LcpLicense.RenewListener,
-            prefersWebPage: Boolean,
-        ): Try<Instant?, LcpError> =
-            Try.failure(failure = LcpError.Unknown(throwable = Exception()))
-
-        override suspend fun returnPublication(): Try<Unit, LcpError> =
-            Try.failure(failure = LcpError.Unknown(throwable = Exception()))
-
-        override suspend fun decrypt(data: ByteArray): Try<ByteArray, LcpError> =
-            Try.success(success = data)
-
-        override fun close() {
-            closed = true
-        }
-
-        override fun canCopy(text: String): Boolean = true
-        override fun canPrint(pageCount: Int): Boolean = true
-        override suspend fun copy(text: String): Boolean = true
-        override suspend fun print(pageCount: Int): Boolean = true
-    }
 
     @Test
     fun `isRestricted is true when license is null`() {
